backend/requirements.txt
alabaster==0.7.8
asgiref==3.5.0
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.16
entrypoints==0.3
feedparser==5.2.1
gitsome==0.8.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.2
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.4.0
pexpect==4.6.0
Pillow==7.0.0
ply==3.11
prompt-toolkit==2.0.10
psycopg==3.0.9
psycopg2-binary==2.9.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
PyYAML==5.3.1
redgreenunittest==0.1.1
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sos==4.2
Sphinx==1.8.5
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
xonsh==0.9.13
zipp==1.0.0
zope.interface==4.7.1
